Add files via upload
The page contains a body/ mass index calculator and the accompanying health status according to the index. To use the tool, launch the html page. A prompt window launches and asks for the body weight and height of the user. After inputting the details and using the calculate tab, the BMI is provided and a health status according to the index is also provided.